directory structure refactor
