+++ README.md
@@ -1,3 +1,3 @@
 # WCC homework
 
-![my gif](public/mygif.gif)
+## WCC homework
