--- src/.env.development
@@ -1,1 +1,0 @@
-REACT_APP_MOVIE_API_KEY=82f71ccc5d36993ad9e836050b46a0ab
