+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { OutputContainer } from './OutputContainer';
-import { SearchContainer } from './SearchContainer';
+import { OutputContainer } from './components/OutputContainer/OutputContainer';
+import { SearchContainer } from './components/SearchContainer/SearchContainer';
 
 const App: React.FC = () => {
   return (
