+++ src/components/Icons/MovieIcon.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
 
-export const MovieIconContainer: React.FC = () => (
+export const MovieIcon: React.FC = () => (
   <svg viewBox="0 0 512 512">
     <path
       d="M352 255.5l-192 96v-192l192 96zm160-224v448H0v-448h512zm-192 64h64v-32h-64v32zm-96 0h64v-32h-64v32zm-96 0h64v-32h-64v32zm-96 0h64v-32H32v32zm64 320H32v32h64v-32zm96 0h-64v32h64v-32zm96 0h-64v32h64v-32zm96 0h-64v32h64v-32zm96 0h-64v32h64v-32zm0-288H32v256h448v-256zm0-64h-64v32h64v-32z"
       fill="currentColor" />
   </svg>
-
 );
