+++ src/constants.ts
@@ -1,0 +1,1 @@
+export const API_KEY = "82f71ccc5d36993ad9e836050b46a0ab";
